Dto parser decomposition
