+++ app_update_checker/lib/config/dto/models/checker_config_dto.dart
@@ -1,4 +1,4 @@
-import '../entity/version.dart';
+import '../../entity/version.dart';
 
 import 'release_dto.dart';
 import 'store_dto.dart';
