+++ app_update_checker/lib/config/dto/models/dto_parser_exception.dart
