+++ app_update_checker/lib/config/dto/models/release_dto.dart
@@ -1,6 +1,6 @@
 import 'dart:ui';
 
-import '../entity/version.dart';
+import '../../entity/version.dart';
 
 import 'store_dto.dart';
