+++ app_update_checker/lib/config/dto/models/store_dto.dart
